+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K83DE39C5TRWV5DNA2NQ8J1W"
+        "01K83MSATHZKHA6MPE6PDBK33K"
     )
     r"""The id of the resource"""